services/trading-bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
redis==5.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
python-dotenv==1.0.1
requests==2.32.3
websockets==13.1
cryptography==43.0.1
pyjwt==2.9.0
python-dateutil==2.9.0.post0
httpx==0.27.2
google-generativeai==0.8.4